# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Disciplr API
  version: 0.1.0
  description: API documentation for Disciplr backend
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Vault:
      type: object
      properties:
        id:
          type: string
          format: uuid
        creator:
          type: string
        amount:
          type: string
        status:
          type: string
          enum:
            - active
            - completed
            - failed
            - cancelled
        startTimestamp:
          type: string
          format: date-time
        endTimestamp:
          type: string
          format: date-time
        successDestination:
          type: string
        failureDestination:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - creator
        - amount
        - status
        - startTimestamp
        - endTimestamp
        - successDestination
        - failureDestination
        - createdAt
    Milestone:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vaultId:
          type: string
          format: uuid
        description:
          type: string
        status:
          type: string
          enum:
            - pending
            - verified
        createdAt:
          type: string
          format: date-time
        verifiedAt:
          type: string
          format: date-time
      required:
        - id
        - vaultId
        - description
        - status
        - createdAt
...
paths:
...
  /api/vaults:
    get:
      summary: List vaults
      tags:
        - Vaults
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Paginated list of vaults
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Vault"
                  pagination: {}
                required:
                  - data
    post:
      summary: Create a new vault
      tags:
        - Vaults
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                creator:
                  type: string
                amount:
                  type: string
                endTimestamp:
                  type: string
                  format: date-time
                successDestination:
                  type: string
                failureDestination:
                  type: string
                milestones:
                  type: array
                  items:
                    type: object
                    properties:
                      description:
                        type: string
                    required:
                      - description
              required:
                - creator
                - amount
                - endTimestamp
                - successDestination
                - failureDestination
      responses:
        "201":
          description: Vault created
          content:
            application/json:
              schema: {}
  /api/vaults/{vaultId}/milestones:
    get:
      summary: Get milestones for a vault
      tags:
        - Milestones
      parameters:
        - name: vaultId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of milestones
          content:
            application/json:
              schema:
                type: object
                properties:
                  milestones:
                    type: array
                    items:
                      $ref: "#/components/schemas/Milestone"
                required:
                  - milestones
...
  /api/analytics/summary:
    get:
      summary: Get analytics summary
      tags:
        - Analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema: {}
...
  /api/analytics/vaults:
    get:
      summary: Get vault analytics
      tags:
        - Analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Vault analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  vaults:
                    type: array
                    items: {}
                  generatedAt:
                    type: string
                    format: date-time
                required:
                  - vaults
                  - generatedAt
  /api/analytics/vaults/{id}:
    get:
      summary: Get vault analytics by ID
      tags:
        - Analytics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vault analytics
  /api/analytics/milestones/trends:
    get:
      summary: Get milestone trends
      tags:
        - Analytics
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: groupBy
          in: query
          schema:
            type: string
            enum:
              - day
              - week
      responses:
        "200":
          description: Milestone trends
          content:
            application/json:
              schema:
                type: object
                properties:
                  from:
                    type: string
                    format: date-time
                  to:
                    type: string
                    format: date-time
                  groupBy:
                    type: string
                    enum:
                      - day
                      - week
                  buckets:
                    type: array
                    items: {}
                required:
                  - from
                  - to
                  - groupBy
                  - buckets
  /api/analytics/behavior:
    get:
      summary: Get user behavior analytics
      tags:
        - Analytics
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
        - name: baseScorePerSuccess
          in: query
          schema:
            type: integer
            default: 10
        - name: penaltyPerFailure
          in: query
          schema:
            type: integer
            default: 5
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: User behavior score
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  successes:
                    type: number
                  failures:
                    type: number
                  behaviorScore:
                    type: number
                  evaluatedFrom:
                    type:
                      - string
                      - "null"
                    format: date-time
                  evaluatedTo:
                    type:
                      - string
                      - "null"
                    format: date-time
                required:
                  - userId
                  - successes
                  - failures
                  - behaviorScore
                  - evaluatedFrom
                  - evaluatedTo
...